pyproject.toml
[project]
name = "joao-santos-ddf-tech-122025"
version = "0.1.0"
description = "Case Técnico Dadosfera"
authors = [
    {name = "Joao Pedro Santos",email = "87459886+JpedroCSantos@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12.8,<3.13"
dependencies = [
    "ruamel-yaml (<0.18)",
    "dotenv (>=0.9.9,<0.10.0)",
    "great-expectations (==0.18.19)",
    "openai (>=2.14.0,<3.0.0)",
    "textblob (>=0.19.0,<0.20.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "google (>=3.0.0,<4.0.0)",
    "gcsfs (>=2025.12.0,<2026.0.0)",
    "google-cloud-bigquery (>=3.39.0,<4.0.0)",
    "pandas-gbq (>=0.32.0,<0.33.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
